+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Dev: Temporarily disable High DPI scaling on Qt6 builds on Windows. (#4767)
 - Dev: Tests now run on Ubuntu 22.04 instead of 20.04 to loosen C++ restrictions in tests. (#4774)
 - Dev: Do a pretty major refactor of the Settings classes. List settings (e.g. highlights) are most heavily modified, and should have an extra eye kept on them. (#4775)
+- Dev: Remove `boost::noncopyable` use & `boost_random` dependency. (#4776)
 
 ## 2.4.5